requirements.txt
selenium
webdriver_manager
trafilatura
livekit-agents[deepgram,openai,silero]>=0.8.0
livekit-plugins-elevenlabs>=0.1.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
livekit-plugins-groq